+++ src/main/java/net/bios/frenchvanilla/Events.java
@@ -1,0 +1,16 @@
+package net.bios.frenchvanilla;
+
+import net.bios.frenchvanilla.armor_switch.UseArmourEvent;
+import net.bios.frenchvanilla.campfire_rest.UseCampfireEvent;
+import net.bios.frenchvanilla.deathlock.PlayerDeathEvents;
+import net.bios.frenchvanilla.deathlock.UseDeathKeyEvent;
+
+public class Events {
+    public static void register() {
+        Commands.registerEvent();
+        PlayerDeathEvents.register();
+        UseDeathKeyEvent.register();
+        UseArmourEvent.register();
+        UseCampfireEvent.register();
+    }
+}