+++ src/main/java/net/bios/frenchvanilla/Settings.java
@@ -12,4 +12,6 @@ public final class Settings extends Config implements ConfigContainer {
 
     public boolean armorSwapping = true;
     public boolean campfireResting = true;
+    public boolean deathLocks = true;
+    public double deathKeyUnlockDistance = 10.0;
 }